requirements.txt
 fastapi
uvicorn
httpx
pydantic
python-dotenv
openai

